TensorImpl: Lazily compute numel and contiguity when symbolic
Currently whenever the sizes or strides are modified for a `TensorImpl` we
eagerly recompute the numel and memory format flags. This is fine for static
shapes as it's all fast C++ code, but for symbolic shapes it runs slow python code.

This instead changes the `SymbolicShapeMeta` object to compute the derived
quantities lazily at the first request. This has the added benefit that we can
now pass assumptions in `empty_tensor_restride` which remove the need to compute
some contiguity flags at all.

ghstack-source-id: 2bf6957107899e6b4a81e1b1a8b7b5977be1b80a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112785